# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Crescendo API",
    "description": "Music artist revenue-share investment platform. Users invest in artists via real revenue-share rights, powered by a double-entry ledger, Chartmetric-driven pricing, and Stripe Connect payouts.\n\n## Auth\nRegister → Login → use the returned JWT as `Authorization: Bearer <token>`.\nAdmin role grants access to all endpoints. Investor/artist roles are scoped.\n\n## Rate Limits\n- Auth endpoints: 20 requests / 15 minutes per IP\n- Trade endpoints: 60 requests / 1 minute per IP\n- All other endpoints: 200 requests / 1 minute per IP\n\nRate limit headers (`RateLimit-*`) are returned on every response (draft-7 standard).\n\n## Idempotency\nTrade buy/sell accept an `Idempotency-Key` header. If the same key is re-sent within 24 hours, the cached response is returned.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT returned from /api/auth/login or /api/auth/register"
      }
    },
    "schemas": {
...
      "Artist": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "stageName": { "type": "string" },
          "bio": { "type": "string", "nullable": true },
          "spotifyArtistId": { "type": "string", "nullable": true },
          "sharesOutstanding": { "type": "integer" },
          "maxShares": { "type": "integer" },
          "revenueSharePct": { "type": "string", "example": "0.10" },
          "basePrice": { "type": "string", "example": "1.0000" },
          "currentPrice": { "type": "string", "example": "1.2500" },
          "currentBid": { "type": "string", "example": "1.1875" },
          "currentAsk": { "type": "string", "example": "1.3125" },
          "circuitBreakerStatus": { "type": "string", "enum": ["closed", "tripped"] }
        }
      },
...
      "RiskControls": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "artistId": { "type": "string", "format": "uuid", "nullable": true },
          "isGlobal": { "type": "boolean" },
          "maxPositionShares": { "type": "integer" },
          "maxPositionPct": { "type": "string" },
          "dailyTradeCapShares": { "type": "integer" },
          "dailyTradeCapUsd": { "type": "string" },
          "cooldownMinutes": { "type": "integer" },
          "circuitBreakerThresholdPct": { "type": "string" },
          "spreadBps": { "type": "integer" }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/admin/run-traction-index": {
      "post": {
        "summary": "Recompute traction index for all artists",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "description": "Admin only. Re-reads latest metric snapshots and recomputes traction scores and prices for every artist.",
        "responses": { "200": { "description": "Computed scores and updated prices per artist" } }
      }
    },
    "/api/admin/traction-index/update": {
      "post": {
        "summary": "Update traction index for one artist (legacy)",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "description": "Admin only. Manually set component scores. If Spotify scores are omitted and the artist has a spotifyArtistId, they are fetched automatically.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["artistId"],
                "properties": {
                  "artistId": { "type": "string", "format": "uuid" },
                  "albumVelocity": { "type": "number", "minimum": 0, "maximum": 100 },
                  "catalogSize": { "type": "number", "minimum": 0, "maximum": 100 },
                  "revenueGrowth": { "type": "number", "minimum": 0, "maximum": 100 },
                  "socialFollowers": { "type": "number", "minimum": 0, "maximum": 100 },
                  "externalPopularity": { "type": "number", "minimum": 0, "maximum": 100 }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated snapshot with new price" } }
      }
    },
    "/api/admin/risk-controls/{artistId}": {
      "put": {
        "summary": "Update risk controls for an artist",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "description": "Admin only. Creates or updates per-artist risk controls. Only provided fields are changed.",
        "parameters": [{ "name": "artistId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxPositionShares": { "type": "integer" },
                  "maxPositionPct": { "type": "number" },
                  "dailyTradeCapShares": { "type": "integer" },
                  "dailyTradeCapUsd": { "type": "number" },
                  "cooldownMinutes": { "type": "integer" },
                  "circuitBreakerThresholdPct": { "type": "number" },
                  "spreadBps": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated controls", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RiskControls" } } } },
          "201": { "description": "Created controls (first time)" }
        }
      }
    },
    "/api/admin/risk-controls": {
      "put": {
        "summary": "Update global risk controls",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "description": "Admin only. Updates the global (default) risk controls. Requires seed to have been run.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "maxPositionShares": { "type": "integer" },
                  "maxPositionPct": { "type": "number" },
                  "dailyTradeCapShares": { "type": "integer" },
                  "dailyTradeCapUsd": { "type": "number" },
                  "cooldownMinutes": { "type": "integer" },
                  "circuitBreakerThresholdPct": { "type": "number" },
                  "spreadBps": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated global controls", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RiskControls" } } } },
          "404": { "description": "Global controls not found — run seed first" }
        }
      }
    },
    "/api/admin/circuit-breaker/{artistId}/reset": {
      "post": {
        "summary": "Reset circuit breaker",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "description": "Admin only. Sets circuit breaker status back to 'closed', re-enabling trading for the artist.",
        "parameters": [{ "name": "artistId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Artist with reset circuit breaker", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } } },
          "404": { "description": "Artist not found" }
        }
      }
    },
    "/api/admin/ledger/integrity": {
      "get": {
        "summary": "Verify debits equal credits",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Integrity report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "integrity": {
                      "type": "object",
                      "properties": {
                        "netBalance": { "type": "string", "description": "Should be '0' if ledger is balanced" },
                        "isBalanced": { "type": "boolean" },
                        "totalEntries": { "type": "integer" }
                      }
                    },
                    "accounts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "accountType": { "type": "string" },
                          "balance": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/demo/reset": {
      "post": {
        "summary": "Reset demo state",
        "description": "Admin only. Wipes all transactional data (orders, positions, candles, dividends, ledger entries), preserves artists and risk controls, resets prices to base, reruns traction index. Idempotent.",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Reset complete with step log",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["reset_complete"] },
                    "steps": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "step": { "type": "string" },
                          "detail": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}